refactor tests: extract createContext helper
